Poprawki w algorytmie HighPass oraz API. Podstawowa funkcjonalność API działa poprawnie dla algorytmu HighPass, jednak na ten moment tylko z poziomu bezpośrednich zapytań (nie przez aplikację webową).

+++ PrzetwarzanieObrazow.Code/Algorithms/HighPassFilter.cs
@@ -45,7 +45,7 @@ public class HighPassFilter : ImageAlgorithm
 				g /= 1;
 				b /= 1;
 				
-				OutputImage.SetPixel(i, j, Color.FromArgb(r, g, b));
+				OutputImage.SetPixel(i, j, Color.FromArgb((byte)r, (byte)g, (byte)b));
 			}
 		}
